+++ R/badge.R
@@ -450,7 +450,8 @@ badge_dependencies <- function(pkg = NULL) {
 ##' @author Scott Chamberlain (badges API), Maëlle Salmon (function)
 badge_cran_checks <- function(pkg = NULL) {
   pkg <- currentPackageName(pkg)
-  badge <- paste0("https://cranchecks.info/badges/summary/", pkg)
+  # badge <- paste0("https://cranchecks.info/badges/summary/", pkg)
+  badge <- paste0("https://badges.cranchecks.info/summary/", pkg, ".svg")
   url <- paste0("https://cran.r-project.org/web/checks/check_results_",
                 pkg, ".html")
   placeholder <- "CRAN checks"
@@ -513,4 +514,30 @@ badge_codefactor <- function(ref = NULL) {
   url <- paste0("https://www.codefactor.io/repository/github/", ref)
   svg <- paste0(url, "/badge")
   paste0("[![CodeFactor](", svg, ")](", url, ")")
+}
+
+##' r-universe badge
+##'
+##' @param pkg Package name. If \code{NULL} (the default),
+##'   the name is determined by the URL field in the DESCRIPTION file.
+##' @param user User name. If \code{NULL} (the default),
+##'   the name is determined by the URL field in the DESCRIPTION file.
+##'
+##' @return badge in markdown syntax
+##' @export
+##' @author Alexander Rossell Hayes
+badge_runiverse <- function(pkg = NULL, user = NULL) {
+  if (is.null(pkg) || is.null(user)) {
+    ref <- currentGitHubRef(pkg)
+    ref <- strsplit(ref, "/")[[1]]
+  }
+
+  user <- user %||% ref[[1]]
+  pkg <- pkg %||% ref[[2]]
+
+  paste0(
+    "[![r-universe status badge]",
+    "(https://", user, ".r-universe.dev/badges/", pkg, ")]",
+    "(https://", user, ".r-universe.dev/ui#package:", pkg, ")"
+  )
 }