+++ NEWS.md
@@ -1,3 +1,8 @@
+# badger 0.2.2
+
++ update cran checks badge URL (2022-11-05, Sat, #45)
++ add r-universe badge (2022-07-20, Wed, #42)
+
 # badger 0.2.1
 
 + comment out test for Altmetric (2022-05-23, Mon)